Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-float main () {
+int main () {
     float num1 ;
     float num2 ;
